requirements.txt
absl-py==0.9.0
astor==0.8.1
astunparse==1.6.3
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
cloudpickle==1.3.0
cycler==0.10.0
future==0.18.2
gast==0.3.3
# get==2019.4.13
google-auth==1.17.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
gym==0.17.2
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.1
joblib==0.15.1
Keras==2.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.1
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.4
# post==2019.4.13
protobuf==3.12.2
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
pyntcloud==0.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
query-string==2019.4.13
# request==2019.4.13
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.11.0
six==1.15.0
sklearn==0.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow-gpu==2.3.0
tensorflow-compression==1.3
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.48.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0